Create LICENSE
This is a simple python calculator that performs basic arithemetic operations: addition , substraction , multiplication , and division. The project is begineer-friendly and built using simple input/output statements, if-else conditions, and aPython functions . I created this project to improve my Python skills and understanding how to upload and manage code on GitHub.